README
This work is unsupported. I do not claim any rights nor do I claim ownership of the work within this package. I did however, CONFIGURE many of the config's and tweaked the code within some of the scripts contained in the server addons. I am leaving this on github just in case we decide to start back up again, or someone else finds it useful.